Avoid error 414 when retrieving subnet cidrs for ListNetworks
Bug 1172537

In order to avoif 414 the list subnet requests will be split in
multiple requests. The total URI len for each of these requests
will be lower than 8K (the default maximum for eventlet.wsgi.server).
The patch tries to submit a single request, and if the URI is too
long an exception is raised before the request is sent over the
wire; the exception handler will split the subnet id list, and
submit the new requests.

This patch does not address the case in which the server is
configured with a maximum URI length different from 8K.

Change-Id: Ia2414cd5374a91d3d12215807037a5d46b836ad6