Make --version option print a more detailed client version.
The python quantum client was printing "quantum 2.0" regardless of
the version of the client.  This seems to be the version of the
API but people expect the version of the client to be printed.

I looked at the implementation of --version in nova and keystone.
The one in keystone worked and was fairly straight-forward
utilizing code from the pbr package.  I added this code to version.py
rather than __init__.py.

Change-Id: I53264b454eb16d5c294ad40ab6494387637950db
Fixes: Bug #1190652