The "arg_value" was not getting checked for empty list
In the _merge_args function it is possible that "arg_value" which is
an instance of a list can be an empty list. We still want to append
"value" which is also an instance of list (passed as part of "_extra_values"
argument) to it. arg_value was not getting checked for empty list.

Fixes bug #1108365

Change-Id: Ief028a3fd1cacefcf1beab408e1117f3e0dd588f