Proxy find_resource_by_name_or_id from neutron client
Change I95a351071e68dfc2d67f3895c87f45ad2221767f gutted quantumclient so
it serves as a proxy for neutronclient, but in the process the static
method find_resource_by_name_or_id was removed. Nova is using
find_resource_by_name_or_id directly so it needs to still exist in
quantumclient and proxy through to neutronclient. This is needed for the
stable branches which are still using python-quantumclient.

Closes-Bug: #1233264

Change-Id: I01b099aa34fedb63967517f288905321816eaf18