Work around for httplib2 retries on timeout
httplib2 automatically retries a request on a timeout even if the request
is a POST. This can lead to orphaned objects being created that we
don't know about. This patch sets httplib2.RETRIES to 1 to avoid this.

Note: httplib2.RETRIES is only available in new versions of httplib2.
See: https://code.google.com/p/httplib2/issues/detail?id=124 for more
details.

Co-Authored-By: Ashwin Raveendran <ashw7n@gmail.com>
Co-Authored-By: Sudheendra Murthy <sudhi.vm@gmail.com

Change-Id: I5648b644bb9df1fd2ea219159794d9d122696af0
Closes-bug: #1267649