Add ability to update certain attributes in a subnet
In testing the ability to update allocation_pools, I found that it was
not possible to do it.  This commits adds the ability to update those
attributes which are allowed in a PUT operation.

Added --enable-dhcp to the list of options.  The reason for adding it
was so that the help message would clearly show that it can be used
with a subnet-update command.  Otherwise, the end user may be confused
thinking there is no way to use subnet-update to enable dhcp for a
subnet where it is disabled.

DocImpact

Closes-Bug: #1324189
Change-Id: I52d05ec2284b6910a33df4a5cb7dc3888737acb6