Correct the bash completion of CLI
Currently, the "neutron help" command in CLI doesn't expose the
"bash-completion" command to users, but actually, the "neutron
bash-completion" command is available. Additionally, there is a "complete"
command in outputs of "neutron help", but this command will prints a wrong
result. This patch adds the "bash-completion" command to commands list of
neutron CLI and removes the "complete" command.

Change-Id: I0f2ec22da7ba36f79197acb41d0621fc726cc0f3
Closes-Bug: #1340647