Revert "Fix CLI support for DVR"
This reverts commit bfec80a776a53b554a8f2d023d2974539edef3ce.

This patch needs to be reverted until we resolve the issue with
patch 109180.

The previous change introduced a situation where we can only
override the default server side setting from the client side
for "--distributed" when the default is False. But we cannot
override when the default is True.

This is a requirement for DVR and at the same time we don't
want to introduce any new changes that would affect any current
plugins. So we will decide on it later and in the mean while I
will revert it.

Partial-Bug: #1346121
Partial-Bug: #1347960
Change-Id: I261f9e232b1f72ed0cb2f0c1ba218facaba9afcb