Replace httpretty with requests_mock
The httpretty library has poor Py3 support, currently the only
working version is 0.8.0, all later minor versions are pinned.

Requests-mock library provides mocking layer for requests package. It is
written in such a way that it is fairly simple to convert existing
httpretty test suites to requests-mock. The library is hosted at Stackforge
and gated as usual OpenStack project.

Change-Id: If7f4baefb53976571efc47dda17a7b543112d555