Make secgroup rules more readable in security-group-show
Previously security_group_rules in security-group-show is
displayed in JSON dump in a format of one line per rule,
but each line is too long and hard to understand rules.
After this commit each rule is displayed as JSON dump with
indentation and we can easily read security group rules.

Partial-Bug: #1153766
Change-Id: I0a0d4afad340aee0346c73cf8881900375978858