Add basic functional tests for client library
We don't have a test coverage on neutronclient with HTTPClient.
CLI tests in our unit tests mock out the HTTP client and
We don't have a test combined with keystoneclient.
This leads to bugs in python bindings like bug 1431449.
Neutronclient has two backend HTTP client (HTTPClient and SessionClient)
so it is better to have functional tests for client library itself.

Change-Id: I069a95d5611371730584cf166ceaae6d8e0f6e05