CRUD for VPN endpoint group API
This contains the create, delete, update, show, and list client APIs for
the new VPN endpoint group API.

The API syntax is:

vpn-endpoint-group-create [--name name] [--description desc] \
    --type {subnet|cidr|vlan} --value <value1> [--value <value2>...]

vpn-endpoint-group-delete <name-or-uuid>

vpn-endpoint-group-show <name-or-uuid>

vpn-endpoint-group-update <name-or-uuid> [--name name] [--description desc]

vpn-endpoint-group-list

Notes:

Using prefix "vpn-" on commands, so that there is no conflict with other
neutron commands.

Note: Using --type and --value, instead of --endpoint_type and --endpoint,
which are already used (and must be globally unique). For example, if latter
re used, see error "Must specify new values to update endpoint_group".

The --value may be repeated multiple times, as a list of endpoints can be
added to a group. There must be at least one.

For --type=subnet, the user may enter the name or UUID of subnet, which
must already exist. Client will translate names to UUIDs.

As shown, the endpoints cannot be modified. A new endpoint group must be
formed, and then can be used by VPN (once support is in place).

When updating, either name or description must be provided (otherwise
there is nothing to update).

Added unit tests for the new API.

DocImpact

Change-Id: Ic696674d97474f8bbf3f895aa21b1585a04dc2fe
Partial-Bug: 1459423
Depends-On: Ia729bd0c6967fa2b8c698495aa360f340b42d98a
Depends-On: I6e10590a988312eafca076a14be38b19e2d44a87