Update path to subunit2html in post_test_hook
Per:

http://lists.openstack.org/pipermail/openstack-dev/2015-August/072982.html

The location of subunit2html changed on the images in the gate
so update the path used in the post_test_hook.

Long-term we should just use what's in devstack-gate.

Closes-Bug: #1491646
(cherry picked from commit d75f79f67c176a38d41fcc5f8794d5956ac65ace)

This also adds a gate_hook to invoke d-g for env prepping. It merely
triggers main devstack-gate script. We need the hook because our new
neutronclient jobs define gate_hook function that points to it, and in
that case devstack-gate does NOT trigger its main script, leaving it up
to the hook.

Depends-on: I8549db52fa5b752202ac7f019b3e1931a697e337
Change-Id: If37743547f739e494501bbc5dc3bb214293919fa