Add flavor argument to loadbalancer v2 create
Builds on cli changes for Neutron flavor framework, which have
now been merged.

The '--flavor' argument takes a flavor name or id, looking up the
id from name if needed.

The loadbalancer v2 service can be configured to use the flavor to
find the operator-selected provider and populate the provider field,
enabling dynamic operator control of which provider is being used on
creation.

Change-Id: I02ed05d56fab4d0f60a8616978970336b4415f2c
Implements: blueprint neutron-flavor-framework