Fix arg order of assertEqual and use assertTrue
assertEqual expects that the arguments provided to it should be
(expected, observed).If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped. This commit fixes the issue in python-neutronclient.

Additionally instead of using assertEqual(True,xxx), assertTrue(xxx) should
be used. The same has been fixed in this patch.

Partial-Bug: #1259292
Related-Bug: #1259292

Change-Id: I83ec45af75f61f8005a60efc8176db0df5cfbe9a