Fixed connection_limit and added UT
Due to a typo, the REST API attribute of connection_limit was
not accessed when creating a new lbaas-V2 listener.
This patch fixes the typo, so that the connection_limit can
be initialized when listener is created. Also, the previous
code did not have UT for connection_limit when the listener
was created. The same has been added here.

Closes-Bug: #1515112

Change-Id: I71eef7b92de91edc1f3ce48f40c645bca704427d