Use raw values when non-table formatter is used
neutron CLI supports various output formats such as json, yaml
(e.g. neutron net-list -f json),
but the current output formattings is optimized to the table
formatter to get human-friendly output in normal CLI use
(For example, new lines are inserted).
However, this kind of formatting is unnecessary when other output
formatters are used and prevents users from getting proper data.

This commit changes to skip format_output_data() method when
output formatters other than 'table' are used.
Now we have proper outputs in JSON, YAML or other formats.

In CreateCommand and ShowCommand, cleanup_output_data() is moved
to each command class because it should be called regardless of
a formatter type.

In ListCommand, value formatter for CSV is still supported for backward
compatiblity, but the value formatter for fixed_ips of the port resource
is removed because it brings no much value in CSV output.

Closes-Bug: #1524624
Change-Id: I2668fa90402d7119db69f09a20fb7c7270c9616e