Allow tenant_id positional in quota syntax
This patch allows tenant_id as a positional argument to the
neutron 'quota-show', 'quota-update', and 'quota-delete' command.
This is helpful because the positional argument is thrown away
by default and the calling tenant ID is used instead so a user
can unexpectedly update or view the wrong quota.

This syntax is still hidden from the help output to encourage the
more explicit --tenant-id syntax.

Closes-Bug: #1524532
Change-Id: I83d4e0763cc8513a5d608a3754f7d46eaa3f1f6b