Show tenant_id in *-list by admin
If API call is done by admin, neutron API returns all resources
from all tenants. We received several times a request that
tenant_id is useful in such case.

This patch displays tenant_id column when *-list is invoked
by a user with 'admin' role by default. If columns to be displayed
are specified by '-c' option, the specified columns are honored
and tenant_id columns is not added.

Add functional tests to cover this change. This feature depends
on keystoneauth1 and unit tests cannot cover this change.

Closes-Bug: #1526238
Change-Id: Id58a0b795f76b4fe24aaf89f1f6486c23f19a553