Fix random failure of security group unit test
CLITestV20SecurityGroupsJSON.test_extend_list_exceed_max_uri_len
fails randomly because this test uses query parameters and
the order of query parameters in a construct URL is not deterministic.
As a result the test failed randomly.
The test should use MyUrlComparator to compare URLs.
MyUrlComparator ignores the order parameters when comparing URLs.

Change-Id: I8a8116f45583102fa1351d17a1cc5174bdce3cc9
Closes-Bug: #1570689