Unable to add classless-static-route in extra_dhcp_opt extension
When adding classless-static-route in extra_dhcp_opt for a port,
neutron client complains syntax error. The reason is neutron client
interprets the "," inside the opt_value as a delimiter of key-value
pairs for --extra-dhcp-opt.

This patch modifies neutron client to allow commas inside opt_value.

This is one of two proposed solutions described in #1605421.

Change-Id: I8ade427885d11f86d594cecaed96fcb82e38798f
Closes-Bug: #1605421