Disable VPNaaS functional tests in the neutronclient gate
This addresses failures for neutronclient's functional job of
advanced services. VPNaaS is no longer maintained and thus it
makes sense to stop gating against it. Hook and tests are kept
for local development and in case things will resurrect back
to life.

This patch also makes extension functional tests conditional to
the presence of the extension under test.

Closes-bug: #1645819

Change-Id: I17ab8accaa943b9ac04893b5e131be0afe0f5429