Do not append .json to sending URL
Both SessionClient and HTTPClient set content-type as application/json,
so there is no need to append ".json" as a format specifier.

This commit also cleans up 'format' attribute in various client class
and unit tests.

Also hide --request-format opiton from the command line help as it is
just ignored. We need to keep the option to keep backward-compatibility.

Closes-Bug: #1678488
Change-Id: I5fc69b9dfed9b7b6350a4aad06417a914a80d090