FWaaS UT: Fake rule should return value returned from API
In FWaaS UT, fwaas.fakes.FirewallRule is expected to return the same
values returned from the API, but it returns 'any' for 'protocol'.
In FWaaS API definition, JSON null (None in python) means 'any',
so it should be None rather than 'any'.

This commit fixes it and related codes in test_firewall.py.
_generate_data() generates expected formatted data from API data.
The same logic is also needed in _update_expect_response(). This commit
introduces a common function _replace_display_columns to do this.

Change-Id: If4fa6b26f72003b6c94020a762a5c19d17b338cd