Populate shell.COMMANDS dict not to break CLI extension UT
shell.COMMANDS is not intended to be public.
However, 8 projects using neutron CLI extenions access this variable
in their UT. I am honestly surprised on this, but it is the real.
Neutron CLI is now deprecated, so it is better not to break them.

This commit populates shell.COMMANDS dict based on registered commands.

The following is my testting results with 'neutron net-list'.
I ran this 10 times and the average user time is as follows:
 before 1.096 sec
 after  1.112 sec

More dynamic loading can be explored but I think
this overhead can be acceptable.

Closes-Bug: #1706573
Change-Id: Idcbc64e6f709a666829e0c2985a03aa4517ccf77