Make func test work with keystone v3 only
neutronclient func test is broken due to missing fully support
of keystone v3 API. It seems keystone v2 API has been disbaled recently
and this triggers the test failures.

There are two causes.

The one is because tempest.lib.cli.base supports only keystone v2 auth info.
There is an on-going review in tempest, but until the next
tempest release we need a workaround. The workaround is to pass
keystone v3 related arguments in neutron() method.
'flags' argument allows us to pass extra auth info to the CLI.

The other reason is because HTTPClient (the legacy python binding class)
supports only keystone v2 API. keystone v3 API is provided via SessionClient.
I believe it is not a good idea to keep the two types of python binding
classes in future. It looks like a time to drop HTTPClient.
As a start, this commit drops HTTPClient func tests.
Note that SessionClient supports keystone v2 as well.

Closes-Bug: #1721553
Related-Bug: #1719687
Change-Id: I6f843e1412400bb1dfb4fc2352fc5cc4c3b837a4