Move grenade neutron-lib job to py3 and in python-neutronclient repo
'legacy-grenade-dsvm-neutron-libs' grenade neutron-lib jobs
is present in opensatck-zuul-jobs repo running on python 2. This needs
to run on py3 from Ussuri onwards and py2 for stable/branch.

python-neutronclient is the only repo which is using this job.

py2 version has been kept in opensatck-zuul-jobs and for ussuri
onwwards this has been migrated to python-neutronclient repo with
py3 version.

Change-Id: I165f86e1d05c8830e0b75b5c2db369178a3df7e7