Add support for port bindings
The port bindings API has been around since Pike but has never been
exposed via neutronclient. We still use this in nova so it would be nice
to avoid having to create a KSA client manually for this stuff. We don't
need to expose things via the UI, just the client itself, so do that.

Change-Id: Ied1a057186f0819166df84725b09ded314930a1d
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Sem-Ver: feature